[move, rename] SliceResponseDto, SortResponseDto로 클래스명 변경 및 global/util 패키지로 이동
related to: #89
